refactor(share-kit): simplify-pass fixes
Applied after a three-agent code review (reuse / quality / efficiency):

- Extract sanitizeFilename into lib/filename.ts; export/index.ts and
  storage/spool-file.ts both now import the single implementation
  instead of carrying near-identical copies.
- Move the truncate(s, n) helper into parsers/source.ts so the
  ChatGPT and Claude extractors share one definition.
- Drop the filenameBase indirection in export/index.ts (it was a
  zero-op wrapper around filenameFor); fold the single call site to
  use filenameFor directly.
- Memoize Body's preprocess output and the markdown components map.
  The redact regex now compiles once per (text, redact) change
  instead of every render; the components map only rebuilds when
  accent / accentBg / blockStroke change.
- Normalize body.tsx indentation that drifted during an earlier
  refactor to a typed components map.
- Move react-markdown, remark-gfm, and lucide-react from
  dependencies to peerDependencies (with widened ranges that cover
  both v9/v10 react-markdown and v0/v1 lucide-react). Hosts already
  declare these, so peer-only avoids double-bundling. Kept as
  devDependencies inside the package so vitest + Vite still resolve
  them for local typecheck and test.
