+++ packages/share-kit/package.json
@@ -28,20 +28,23 @@
   },
   "dependencies": {
     "html-to-image": "^1.11.13",
-    "idb-keyval": "^6.2.1",
-    "lucide-react": "^0.460.0",
-    "react-markdown": "^10.1.0",
-    "remark-gfm": "^4.0.1"
+    "idb-keyval": "^6.2.1"
   },
   "peerDependencies": {
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "lucide-react": "^0.460.0 || ^1.0.0",
+    "react-markdown": "^9.0.0 || ^10.0.0",
+    "remark-gfm": "^4.0.0"
   },
   "devDependencies": {
     "@types/node": "^22.19.17",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "@vitejs/plugin-react": "^4.7.0",
+    "lucide-react": "^0.460.0",
+    "react-markdown": "^10.1.0",
+    "remark-gfm": "^4.0.1",
     "vite": "^6.3.3",
     "vite-plugin-dts": "^4.5.0",
     "vitest": "^3.2.4"
